# OpenAPI.yaml
openapi: 3.0.0
info:
  title: store
  version: '1.0'
  description: A simple store
  contact:
    name: SecuritEase Dev
    url: 'https://www.securitease.com'
    email: internal@securitease.com
  termsOfService: 'https://www.securitease.com'
  license:
    name: Apache 2.0
    url: 'https://www.apache.org/licenses/LICENSE-2.0.htm'
servers:
  - url: 'http://localhost:8080'
    description: dev
paths:
  /order:
    get:
      summary: Get Orders
      tags:
        - store
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
      operationId: getOrders
      description: Get all orders
    post:
      summary: Create Order
      tags:
        - store
      responses:
        '201':
          description: Created
      operationId: createOrder
      description: Create a new order
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                description:
                  type: string
                customerId:
                  type: integer
                  format: int64
  /order/{id}:
    get:
      summary: Get Order by ID
      tags:
        - store
      parameters:
        - in: path
          name: id
          schema:
            type: integer
            format: int64
          required: true
          description: Order ID
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
      operationId: getOrderById
      description: Get an order by ID
  /customer:
    get:
      summary: Get Customers
      tags:
        - store
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Customer'
      operationId: getCustomers
      description: Get all customers
    post:
      summary: Create Customer
      tags:
        - store
      responses:
        '201':
          description: Created
      operationId: createCustomer
      description: Create a new customer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
  /customer/{id}:
    get:
      summary: Get Customer by ID
      tags:
        - store
      parameters:
        - in: path
          name: id
          schema:
            type: integer
            format: int64
          required: true
          description: Customer ID
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
      operationId: getCustomerById
      description: Get a customer by ID
  /products:
    get:
      summary: Get Products
      tags:
        - store
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
      operationId: getProducts
      description: Get all products
    post:
      summary: Create Product
      tags:
        - store
      responses:
        '201':
          description: Created
      operationId: createProduct
      description: Create a new product
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                description:
                  type: string
  /products/by-orders:
    get:
      summary: Get Product IDs by Order IDs
      tags:
        - store
      parameters:
        - in: query
          name: orderIds
          schema:
            type: array
            items:
              type: integer
              format: int64
          required: true
          description: Order IDs
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: integer
                  format: int64
      operationId: getProductIdsByOrderIds
      description: Get product IDs by order IDs
  /customers/search:
    get:
      summary: Search Customers
      tags:
        - store
      parameters:
        - in: query
          name: query
          schema:
            type: string
          required: true
          description: Search query
        - in: query
          name: page
          schema:
            type: integer
          required: false
          description: Page number
        - in: query
          name: size
          schema:
            type: integer
          required: false
          description: Page size
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/Customer'
                  totalElements:
                    type: integer
      operationId: searchCustomers
      description: Search customers by name
components:
  schemas:
    Order:
      title: Order
      type: object
      properties:
        id:
          type: integer
          format: int64
        description:
          type: string
        customer:
          $ref: '#/components/schemas/Customer'
    Customer:
      title: Customer
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        orders:
          type: array
          items:
            $ref: '#/components/schemas/Order'
    Product:
      title: Product
      type: object
      properties:
        id:
          type: integer
          format: int64
        description:
          type: string
tags:
  - name: store
    description: store operations